task completed

+++ src/test/java/com/workintech/s18d1/ApplicationPropertiesAndControllerTest.java
@@ -142,7 +142,7 @@ class ApplicationPropertiesAndControllerTest {
     @Test
     @DisplayName("Find burger by id test")
     void testGetBurgerById() throws Exception {
-        given(burgerDao.findById(sampleBurger.getId())).willReturn(sampleBurger);
+        given(burgerDao.findById((long) sampleBurger.getId())).willReturn(sampleBurger);
 
         mockMvc.perform(get("/burger/{id}", sampleBurger.getId()))
                 .andExpect(status().isOk())
@@ -168,7 +168,7 @@ class ApplicationPropertiesAndControllerTest {
     @DisplayName("Remove burger test")
     void testRemoveBurger() throws Exception {
 
-        given(burgerDao.remove(sampleBurger.getId())).willReturn(sampleBurger);
+        given(burgerDao.remove((long) sampleBurger.getId())).willReturn(sampleBurger);
 
         mockMvc.perform(delete("/burger/{id}", sampleBurger.getId()))
                 .andExpect(status().isOk());
@@ -191,9 +191,9 @@ class ApplicationPropertiesAndControllerTest {
     @DisplayName("Find by price test")
     void testFindByPrice() throws Exception {
         List<Burger> burgers = Arrays.asList(sampleBurger);
-        given(burgerDao.findByPrice(sampleBurger.getPrice().intValue())).willReturn(burgers);
+        given(burgerDao.findByPrice(sampleBurger.getPrice())).willReturn(burgers);
 
-        mockMvc.perform(get("/burger/price/{price}", sampleBurger.getPrice().intValue()))
+        mockMvc.perform(get("/burger/price/{price}", sampleBurger.getPrice()))
                 .andExpect(status().isOk())
                 .andExpect(jsonPath("$", hasSize(1)))
                 .andExpect(jsonPath("$[0].name", is(sampleBurger.getName())));
@@ -211,5 +211,3 @@ class ApplicationPropertiesAndControllerTest {
                 .andExpect(jsonPath("$[0].contents", containsString("Cheese")));
     }
 }
-
-
